1. applied Vaclav's patches for gcc 3.0 compilation 2. blind fix for assert failure in MF dtor for POP3 3. added support for logging into a file

+++ src/adb/ImportText.cpp
@@ -207,7 +207,7 @@ bool AdbTextImporter::TestDelimiter(char chDelimiter)
       size_t maxDiff = nDelimitersAverage / 10;
       for ( nLine = 0; m_lastTestResult && (nLine < nLines); nLine++ )
       {
-         if ( abs(nDelimiters[nLine] - nDelimitersAverage) > (int)maxDiff )
+         if ( abs((int)(nDelimiters[nLine] - nDelimitersAverage)) > (int)maxDiff )
          {
             // this line looks strange...
             return FALSE;
